Trait obce::codegen::MethodDescription

pub trait MethodDescription<const METHOD_HASH: u32> {
    type Input;
    type Output;

    const ID: u16;
}

Chain extension method description.

METHOD_HASH generic is dependent solely on the method name, while ID can be changed via #[obce(id = ...)] macro.

Required Associated Types

type Input

Method input type, that is required for chain extension calls.

type Output

Method output type, that you can use to obtain results from chain extension calls.

Required Associated Constants

const ID: u16

Unique chain extension method identifier.