Trait obce::ink_lang::storage::traits::Storable

pub trait Storable: Sized {
    // Required methods
    fn encode<T>(&self, dest: &mut T)
       where T: Output + ?Sized;
    fn decode<I>(input: &mut I) -> Result<Self, Error>
       where I: Input;
}

Trait for representing types which can be read and written to storage.

This trait is not the same as the scale::Codec. Each type that implements scale::Codec are storable by default and transferable between contracts. But not each storable type is transferable.

Required Methods

fn encode<T>(&self, dest: &mut T)where T: Output + ?Sized,

Convert self to a slice and append it to the destination.

fn decode<I>(input: &mut I) -> Result<Self, Error>where I: Input,

Attempt to deserialize the value from input.